Refactor code
Move update case data logic from CcdCasesSender to separate class
because the class is too big and it is doing many different things

Fix tests
